+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # CHANGELOG
+
+## 2019.01.19-19:00
+
++ `HashTable` class implemented fully.
++ Helper class `PrimeNumbers`.
++ Helper class `Pair` for a key-value pair object.